+++ src/components/common/Tabs/TabList.tsx
@@ -1,11 +1,8 @@
-import type { ReactElement, ReactNode } from 'react'
+import type { ReactElement } from 'react'
 import { cloneElement } from 'react'
+import type { TabListProps } from './types'
 import { childrenMap } from '@utils'
 
-type TabListProps = {
-  children: ReactNode
-  className?: string
-}
 export const TabList = ({
   children,
   className
