+++ src/components/common/Tabs/TabPanels.tsx
@@ -1,11 +1,8 @@
-import type { ReactElement, ReactNode } from 'react'
+import type { ReactElement } from 'react'
 import { cloneElement } from 'react'
+import type { TabPanelsProps } from './types'
 import { childrenMap } from '@utils'
 
-type TabPanelsProps = {
-  children: ReactNode
-  className?: string
-}
 export const TabPanels = ({
   children,
   className
