+++ src/components/common/Tabs/Tabs.test.tsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event'
 import type { ReactElement } from 'react'
 import { Tab } from './Tab'
 import { Tabs } from './Tabs'
-import type { TabsProps } from './Tabs'
+import type { TabsProps } from './types'
 
 const DummyTabs = ({
   defaultTabIndex
