+++ src/components/common/Tabs/Tabs.tsx
@@ -1,16 +1,11 @@
-import type { ReactElement, ReactNode } from 'react'
+import type { ReactElement } from 'react'
 import { useEffect, useMemo, useState } from 'react'
 import { TabList } from './TabList'
 import { TabPanel } from './TabPanel'
 import { TabPanels } from './TabPanels'
 import { TabsActionContext } from './TabsActionContext'
+import type { TabsProps } from './types'
 
-export type TabsProps = {
-  children: ReactNode
-  defaultTabIndex?: number
-  className?: string
-  onChange?(prevIndex?: number, currentIndex?: number): void
-}
 export const Tabs = ({
   children,
   onChange,
